Code/requirements.txt
tinydb
paho-mqtt
matplotlib
scikit-learn
pandas
numpy
dash
plotly
